extract some function from build_remediations.py to external file
functions and data structures are used in more modules
especially when building remediations for CPE items
